Update college-notes.md
The wiki-graph didn't work right, which resulted in my reducing the intensity of my efforts to determine proper wiki-linking throughout the archive of notes.  I'd still like to see the wiki-graph work someday with this software, but I've adjusted my workflow so that it's no longer a priority to me.