Update metr101ch5-notes.md
Other people experience a stabler world than I do, and that annoys me.  The map around my hometown seems bizarrely different than it was when I grew up there as well, and in ways that normal development can't account for...  I do my best to study the causal systems I'm presented to and return value in thought, but I feel like I'm engaging scattershot with a local region of the multiverse instead of getting to live in one stable reality.  :/