Update envs100disc4-pandemicblues.md
Not sure how to proceed with this draft.  The task of digging up the citations is mechanically straightforward.  If someone else had written this with the same citation gaps, I could easily look them up, fill them in, and consider it good work.  Yet it's my own, and that hurts.