Update envs100-nw-state-parks-essay.md
I know what the problem is.  The environmental courses are /environmentalist/ courses.  They have obligatory answers that everyone already knows.  These answers are presented as scientific reality, and that's /probably/ true.  Indeed, it's /presumably/ true.

Yet it's a little hard to work around the bits of environmentalism that I think might be wrong, because the body of answers is already presumed to be complete.