create 2021-2-4-Twenty-Second.md
I'm really, desperately not supposed to be a person who would be a competent addition to a laboratory, but I really actually am a person who would be a competent addition to a laboratory.

And ironically, my mother's history was similar to my own, in that she was sexually exploited for years before going into biochemistry.