Update envs100disc7-climatejustice.md
There's a necessary acceleration which kicks humanity into seeking nth-order goods rather than breeding in desperation, and compassionate efficiency along with *everything else that makes life worth living* is an nth-order good.  This goes back to eco-industrialism for me.  There are all these societies on Earth that never got *the kick* and still exist in a sociopathic ancestral mode.