Update envs100disc7-climatejustice.md
If I'd noticed how much of a challenge this assignment would be, I might've enjoyed it.  On the other hand, I thought the same thing about the Economics and Weather topic... and I was kind of sick this last week.  I don't know that I would've enjoyed it at all.  I would've liked to remember the Economics and Weather topic, but working on this one might've just made me a lot more stressed about feeling unwell.