Update 2021-2-27-Forty-First.md
More geo-engineering frustrations.  My experiments are not outside of humanity's wealth budget and would not take billions of dollars.  The stuff I want to do is not less expensive than filling a large car dealership with cars, but it also shouldn't be MORE expensive, because the skills and efforts involved are analogous.  If only people believed new things were possible!