update notes-on-tolerogenic-dendritic-cells.md
Part of what's stalling me here is that I achieved some idea of what I'd like to do... and then I didn't really engage with it at the level of creating potential experimental protocols.  Sigh.

Which, to be fair, is also the whole deficit in my skills that I wanted to fix with this project.