update ideological-interpretations-of-The-Viceroy.md
I thought more about the jaunting grids.  This technology mystified me, making me reluctant to implement it, but curious what it could mean for society if it were implemented.  I now think it's a very high-smoothness technology, though not necessarily a good idea save in the presence of other very advanced technologies.