long post contemplating sexual art
I'm well aware 'not supposed to' write about this material, but I chafe at that restriction just as I chafed at the earlier obligation to write about this material.  Contemplation of sexual subjects should by means of concealed coercion be neither forced to exist nor forced to not exist, but in this matter freedom of conscience should reign in whole.