Update 2021-8-07-One-Hundred-and-Sixtieth.md
I'm supposed to fail, because I don't obey rules about who is and isn't allowed to be anti-authoritarian.  I was taught anti-authoritarianism meant sexual obedience.  I don't think that was right.  I think that was actually authoritarian, and that the all-seeing set in society had no right to punish me for rejecting the provision of sexual services.