Update ideological-interpretations-of-The-Viceroy.md
Added conversion notes.  Plus, I lost something petty that I needed for research, and when I found it again I thought of the Potentialities Overseer.  That loss was a petty inefficiency.  Installing an AI mind to watch for every petty inefficiency would surely not pay off... but it'd do stuff day-to-day to keep systems functioning nonetheless.